+++ rs-application/application-api-next/src/lib.rs
@@ -32,24 +32,24 @@ impl App {
     }
 
     fn router() -> Router {
-        Router::new()
-            .nest("/api/v1", routes::api_v1())
-            .route("/health", get(|| async { "success" }))
-            .fallback(|| async {
-                Response::<String>::new(Some(404), Some("Not Found".to_string()), None)
-            })
-            .layer(
-                ServiceBuilder::new()
-                    .layer(SetRequestIdLayer::x_request_id(MakeRequestUuid))
-                    .layer(
-                        TraceLayer::new_for_http()
-                            .make_span_with(RequestIdMakeSpan)
-                            .on_failure(OnFailureBehaviour),
-                    )
-                    .layer(axum::middleware::from_fn(middleware::log_response))
-                    .layer(axum::middleware::from_fn(middleware::log_request))
-                    .layer(PropagateRequestIdLayer::x_request_id())
-                    .layer(CorsLayer::permissive()),
-            )
+        // Router::new()
+        //     .nest("/api/v1", routes::api_v1())
+        //     .route("/health", get(|| async { "success" }))
+        //     .fallback(|| async {
+        //         Response::<String>::new(Some(404), Some("Not Found".to_string()), None)
+        //     })
+        //     .layer(
+        //         ServiceBuilder::new()
+        //             .layer(SetRequestIdLayer::x_request_id(MakeRequestUuid))
+        //             .layer(
+        //                 TraceLayer::new_for_http()
+        //                     .make_span_with(RequestIdMakeSpan)
+        //                     .on_failure(OnFailureBehaviour),
+        //             )
+        //             .layer(axum::middleware::from_fn(middleware::log_response))
+        //             .layer(axum::middleware::from_fn(middleware::log_request))
+        //             .layer(PropagateRequestIdLayer::x_request_id())
+        //             .layer(CorsLayer::permissive()),
+        //     )
     }
 }