+++ huawei/atomicservice/MFA/CHANGELOG.md
@@ -1,3 +1,37 @@
+## v1.3.0
+
+### added
+
+- feat: 支持明亮模式和暗黑模式自动切换(#113)
+
+### optimized
+
+- optimize: 优化颜色相关代码定义，使用主题方式替换(#113)
+
+## v1.2.1
+
+### optimized
+
+- optimize: 优化各个编辑页面返回时的效果(#110)
+- optimize: 优化登陆提示文案(#112)
+
+## v1.2.0
+
+### added
+
+- feat: 「我的」页面增加(#108)
+- feat: 「首页」增加手动输入方式(#108)
+- feat: 「首页」增加下拉刷新功能(#108)
+
+### optimized
+
+- optimize: 优化各类操作体验与提示文案(#108)
+- optimize: 使用 api20 新特性优化首页导航 navigation 逻辑(#108)
+
+### changed
+
+- change(internal): 优化 string 常量命名(#108)
+
 ## v1.1.0
 
 ### added